// src/types/types.ts
export interface Member {
  id: string;
  name: string;
  dataNascimento: string; // Data de Nascimento
  telefone: string; // Telefone para Emergência
  contatoEmergencia: string; // Contato de Emergência
  endereco: string; // Endereço
  alergias: string; // Alergias
  tipoSanguineo: string; // Tipo Sanguíneo
  condicaoMedica: string; // Condição Médica
  medicacao: string; // Medicação
  observacao: string; // Observação
  dataUltimaAtualizacao: string; // Data da Última Atualização
}